pyproject.toml
[project]
name = "fyp-archive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.1.0",
    "ale-py==0.10.1",
    "anyio==4.8.0",
    "appnope==0.1.4",
    "argon2-cffi-bindings==21.2.0",
    "argon2-cffi==23.1.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.4",
    "attrs==24.3.0",
    "babel==2.16.0",
    "beautifulsoup4==4.12.3",
    "bleach==6.2.0",
    "box2d-py==2.3.5",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "cloudpickle==3.1.0",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "debugpy==1.8.11",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "dotenv==0.9.9",
    "etils==1.11.0",
    "exceptiongroup==1.2.2",
    "executing==2.1.0",
    "farama-notifications==0.0.4",
    "fastjsonschema==2.21.1",
    "filelock==3.16.1",
    "fonttools==4.55.3",
    "fqdn==1.5.1",
    "fsspec==2024.12.0",
    "glfw==2.8.0",
    "gym==0.26.2",
    "gym-notices==0.0.8",
    "gymnasium==1.0.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "imageio==2.36.1",
    "importlib-resources==6.5.2",
    "ipykernel==6.29.5",
    "ipython==8.31.0",
    "ipywidgets==8.1.5",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "json-tricks==3.17.3",
    "json5==0.10.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter==1.1.1",
    "jupyter-client==8.6.3",
    "jupyter-console==6.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.11.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.15.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.3.4",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.13",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.0",
    "mpmath==1.3.0",
    "mujoco==3.2.6",
    "nbclient==0.10.2",
    "nbconvert==7.16.5",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "notebook==7.3.2",
    "notebook-shim==0.2.4",
    "numpy==2.2.1",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.1.0",
    "platformdirs==4.3.6",
    "prometheus-client==0.21.1",
    "prompt-toolkit==3.0.48",
    "psutil==6.1.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pygame==2.6.1",
    "pygments==2.19.1",
    "pyopengl==3.1.7",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-json-logger==3.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "readchar==4.2.1",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rpds-py==0.22.3",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "setuptools==75.8.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "stack-data==0.6.3",
    "swig==4.3.0",
    "sympy==1.13.1",
    "terminado==0.18.1",
    "tinycss2==1.4.0",
    "tomli==2.2.1",
    "torch==2.5.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "types-python-dateutil==2.9.0.20241206",
    "typing-extensions==4.12.2",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "widgetsnbextension==4.0.13",
    "zipp==3.21.0",
]
